# docs/api/moodly-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moodly API",
    "description": "Moodly backend API. Authenticate requests with an Auth0 access token.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/entries/today/mood": {
      "put": {
        "tags": [
          "Daily entries"
        ],
        "summary": "Set today's mood",
        "description": "Sets the mood in the authenticated user's entry for today.",
        "operationId": "setTodayMood",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMoodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDailyEntry"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/me/avatar/upload-signature": {
      "post": {
        "tags": [
          "Avatar"
        ],
        "summary": "Create an avatar upload signature",
        "description": "Creates a signed upload payload for Cloudinary after validating the image content type and size.",
        "operationId": "signature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUploadSignature"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/me/avatar/confirm": {
      "post": {
        "tags": [
          "Avatar"
        ],
        "summary": "Confirm an uploaded avatar",
        "description": "Verifies the uploaded Cloudinary asset and saves it as the current user's avatar.",
        "operationId": "confirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAvatar"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/internal/cdc/reindex": {
      "post": {
        "tags": [
          "CDC maintenance"
        ],
        "summary": "Reindex daily entries",
        "description": "Rebuilds the daily-entry search index. Requires the X-Maintenance-Key header.",
        "operationId": "reindex",
        "parameters": [
          {
            "name": "X-Maintenance-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReindexResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/internal/cdc/dead-letters/{id}/replay": {
      "post": {
        "tags": [
          "CDC maintenance"
        ],
        "summary": "Replay a CDC dead letter",
        "description": "Replays a failed CDC delivery by ID. Requires the X-Maintenance-Key header.",
        "operationId": "replay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Maintenance-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "CDC dead letter replayed successfully."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/entries/today": {
      "patch": {
        "tags": [
          "Daily entries"
        ],
        "summary": "Update today's habit log",
        "description": "Marks or unmarks a habit in the authenticated user's entry for today.",
        "operationId": "updateTodayHabit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHabitLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDailyEntry"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/me/avatar": {
      "get": {
        "tags": [
          "Avatar"
        ],
        "summary": "Get the current avatar",
        "description": "Returns the authenticated user's current avatar metadata.",
        "operationId": "current",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAvatar"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/entries": {
      "get": {
        "tags": [
          "Daily entries"
        ],
        "summary": "List entries by date range",
        "description": "Returns the authenticated user's daily entries from `from` through `to`, inclusive. Future dates are not allowed.",
        "operationId": "findBetween",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListDailyEntry"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SetMoodRequest": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32",
            "description": "Mood score, from 1 (lowest) to 5 (highest)",
            "example": 4,
            "maximum": 5,
            "minimum": 1
          },
          "tags": {
            "type": "array",
            "description": "Optional labels describing the mood",
            "example": [
              "productive",
              "calm"
            ],
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": "string",
            "description": "Optional note for the day",
            "example": "Finished the release plan."
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorResponse"
            }
          }
        }
      },
      "ApiResponseDailyEntry": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DailyEntry"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DailyEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "mood": {
            "$ref": "#/components/schemas/Mood"
          },
          "habits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HabitLog"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FieldErrorResponse": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "HabitLog": {
        "type": "object",
        "properties": {
          "habitId": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "Mood": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "UploadRequest": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "minLength": 1
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "contentType"
        ]
      },
      "ApiResponseUploadSignature": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UploadSignature"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UploadSignature": {
        "type": "object",
        "properties": {
          "cloudName": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "uploadPreset": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "signature": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string"
          }
        }
      },
      "ConfirmRequest": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "minLength": 1
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "publicId"
        ]
      },
      "ApiResponseAvatar": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Avatar"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Avatar": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "deliveryUrl": {
            "type": "string"
          }
        }
      },
      "ApiResponseReindexResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ReindexResult"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReindexResult": {
        "type": "object",
        "properties": {
          "indexedEntries": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateHabitLogRequest": {
        "type": "object",
        "properties": {
          "habitId": {
            "type": "string",
            "description": "ID of the habit to update",
            "example": "66d34a8e7eb1e840cb8a85c1",
            "minLength": 1
          },
          "done": {
            "type": "boolean",
            "description": "Whether the habit was completed today",
            "example": true
          },
          "note": {
            "type": "string",
            "description": "Optional completion note",
            "example": "Walked for 30 minutes."
          }
        },
        "required": [
          "habitId"
        ]
      },
...
      "ApiResponseListDailyEntry": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyEntry"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-12T10:00:00Z"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseApiError"
            },
            "examples": {
              "default": {
                "value": "{\"success\":false,\"data\":null,\"error\":{\"status\":400,\"code\":\"VALIDATION_FAILED\",\"message\":\"One or more fields are invalid.\",\"path\":\"/example\",\"errors\":[]},\"timestamp\":\"2026-09-12T10:00:00Z\"}\n"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required or the access token is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseApiError"
            },
            "examples": {
              "default": {
                "value": "{\"success\":false,\"data\":null,\"error\":{\"status\":401,\"code\":\"UNAUTHORIZED\",\"message\":\"Authentication is required.\",\"path\":\"/example\",\"errors\":[]},\"timestamp\":\"2026-09-12T10:00:00Z\"}\n"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user is not allowed to perform this operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseApiError"
            },
            "examples": {
              "default": {
                "value": "{\"success\":false,\"data\":null,\"error\":{\"status\":403,\"code\":\"FORBIDDEN\",\"message\":\"Access is denied.\",\"path\":\"/example\",\"errors\":[]},\"timestamp\":\"2026-09-12T10:00:00Z\"}\n"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseApiError"
            },
            "examples": {
              "default": {
                "value": "{\"success\":false,\"data\":null,\"error\":{\"status\":404,\"code\":\"NOT_FOUND\",\"message\":\"The requested resource was not found.\",\"path\":\"/example\",\"errors\":[]},\"timestamp\":\"2026-09-12T10:00:00Z\"}\n"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponseApiError"
            },
            "examples": {
              "default": {
                "value": "{\"success\":false,\"data\":null,\"error\":{\"status\":500,\"code\":\"INTERNAL_SERVER_ERROR\",\"message\":\"An unexpected error occurred.\",\"path\":\"/example\",\"errors\":[]},\"timestamp\":\"2026-09-12T10:00:00Z\"}\n"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}